Refactor charge data fetching to use a dedicated API function and remove inline logic
